refactor: Extract AudioSourceFactory from AudioManager
Move 5 source creation methods and 12 DI dependencies into a dedicated
AudioSourceFactory, reducing AudioManager constructor from 20 to 9
parameters and the file from 882 to 743 lines. Remove unused
IAudioManager.DeviceManager property from the interface.
